requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
requests
gtts
python-dotenv 